Extracted VueX store for configurations into separate module
